Moved LiteLLM model test to test_llms
